pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "yamflow"
version = "0.1.0"
description = "AI Workflow Framework"
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0",
    "aiohttp>=3.8",
    "click>=8.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=4.1.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src"

[tool.coverage.run]
source = ["src"]